feat(odyssey-eslint): add eslint package
