+++ .eslintrc.js
@@ -1,29 +1,13 @@
-const { pattern, template, header } = require('./header');
+/*!
+ * Copyright (c) 2021-present, Okta, Inc. and/or its affiliates. All rights reserved.
+ * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
+ *
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *
+ * See the License for the specific language governing permissions and limitations under the License.
+ */
 
-module.exports = {
-  parserOptions: {
-    ecmaVersion: 2018,
-    sourceType: 'module'
-  },
-  overrides: [
-    {
-      files: '*.@(js|jsx|ts|tsx)',
-      excludedFiles: [".eslintrc.js", "header.js"],
-      plugins: [
-        'header'
-      ],
-      rules: {
-        'header/header': [
-          'error',
-          'block',
-          [
-            '!',
-            { pattern, template },
-            ...header.split('\n')
-          ],
-          2,
-        ]
-      }
-    }
-  ]
-};
+module.exports = require("@okta/odyssey-eslint");
