+++ package.json
@@ -18,15 +18,13 @@
     "prelerna-publish": "lerna run build",
     "lerna-publish": "lerna publish from-package --no-push --force-publish --no-verify-access --no-verify-registry",
     "lint": "yarn eslint && yarn stylelint",
-    "eslint": "eslint .",
+    "eslint": "eslint . --ext .js,.jsx,.ts,.tsx",
     "stylelint": "stylelint **/*.scss",
     "test": "lerna run test",
     "commitlint": "commitlint --verbose --edit",
     "prepare": "[ -z \"$CI\" ] && husky install || exit 0"
   },
   "devDependencies": {
-    "eslint": "^7.9.0",
-    "eslint-plugin-header": "^3.1.1",
     "lerna": "3.3.2"
   },
   "version": "0.5.0"
