--- packages/odyssey-react/.eslintrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  "env": {
-    "jest": true
-  },
-  parser: '@typescript-eslint/parser',
-  extends: ['plugin:@typescript-eslint/recommended']
-}
