+++ packages/odyssey-react/package.json
@@ -25,11 +25,8 @@
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.7",
     "@types/react-dom": "^17.0.5",
-    "@typescript-eslint/eslint-plugin": "^4.24.0",
-    "@typescript-eslint/parser": "^4.24.0",
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.2",
-    "eslint": "^7.26.0",
     "jest": "^26.6.3",
     "typescript": "^4.2.4"
   },
